split remote and server launches
