--- launch/joy_control.launch.py
@@ -1,29 +1,0 @@
-from launch import LaunchDescription
-from launch_ros.actions import Node
-
-import os
-from ament_index_python.packages import get_package_share_directory
-
-
-def generate_launch_description():
-
-    joy_params = os.path.join(get_package_share_directory('lance_sim'),'config','xbox_controller.yaml')
-
-    joy_node = Node(
-        package='joy',
-        executable='joy_node',
-        parameters=[joy_params],
-    )
-
-    teleop_node = Node(
-        name = 'teleop_node',
-        package='teleop_twist_joy',
-        executable='teleop_node',
-        parameters=[joy_params],
-        remappings=[('/cmd_vel', '/joystick_cmd_vel')]
-    )
-
-    return LaunchDescription([
-        joy_node,
-        teleop_node
-    ])
