+++ launch/robot_state_publisher.launch.py
@@ -16,9 +16,6 @@ def generate_launch_description():
 
 	use_sim_time = LaunchConfiguration('use_sim_time', default = 'false')
 
-	# with open(urdf_path, 'r') as infp:
-	# 	robot_desc = infp.read()
-
 	return LaunchDescription(
 		[
 			DeclareLaunchArgument(
